Redo how hepmc flag/filenames are handled in submit_mc_jobs_htcondor.py.
Now, if the `--hepmc` flag is not used, no HepMC will be produced, as per user's wishes.
If it is used, with no accompanying option, then a filename is autogenerated, and there the files is added to the copyFromLocal list.
If there is an accompanying filename, then "_seedX" is appended to the stem, and again it is queued to the copyFromLocal list.